Support PostgresQL and SQLite backends (#8)
* Handle invalid edges in freeport/backend (#6)

* Reusable iden in mutate/edge.rs

* Cargo fmt

* Update SeaQuery versions in dependency tree (with forked SeaORM and SeaSchema)

* Connectivity: on conflict update

* Do not handle edge collision in Starfish

* Filter out edges that do not satisfy foreign key constraint in freeport/backend

* Log invalid edges filtered out

* Handle empty nodes/edges

* Removed unused imports and cargo fmt

* On conflict update edges

* gitignore SQLite db files

* Enable all `sqlx-*` features for now

* Create db indexes in standalone statement

* Add `sqlx-*` features and propagate to enable the corresponding feature in `sea-orm`

* Fix bug where no error is thrown even though `upsert` is false and node insertion collision occurs

* Update tests/common/setup.rs to initialize testing db env for PostgreSQL & SQLite

* Propagate db backend features from starfish/ to starfish-core/

* Break up query unit tests

* Change all u64 to u32 as postgres doesn't support the former

* Support SQLite in testing

* Add sh script to run cargo test with the specific db backend

* Use sqlx-mysql feature in MySQL, MariaDB CI tests

* Use sqlx-mysql feature in compile-mysql CI task

* Add compile-sqlite and compile-postgres CI tasks

* Add sqlite and postgres CI tasks

* Fix for rocket's updated API

* Use `with-json` feature of sea-orm

* Update postgres database url

* Edit README

* Bump sea-orm version

* Check for invalid edges with all previously stored nodes

* Edit README

* Store crate names to correct data path

* Batch update connectivity for SQLite

* Check if DbBackend is Sqlite to decide whether to batch update connectivity

+++ starfish/starfish-core/src/lang/query.rs
@@ -117,7 +117,7 @@ pub enum QueryConstraintSortByKeyJson {
 pub enum QueryGraphConstraintLimitJson {
     /// Recurse to a certain depth, 0 means root only.
     /// A `null` value means there is no limit
-    Depth(Option<u64>),
+    Depth(Option<u32>),
     /// Include up to this number of nodes in each batch
     /// A `null` value means there is no limit
     BatchSize(Option<usize>),
